Don't hardcode architecture
libyara might need to be build on ARM64 (`aarch64`) architecture too, e.g. I'm using a Mac with an M1 chip, and `node-yara` builds fail on my [dev-and-test environment](https://github.com/Automattic/jetpack-backups-dev-and-test-env-demo) (as the containers run on ARM64 as well.

This sets the [GYP variable](https://gyp.gsrc.io/docs/InputFormatReference.md#command-expansions) to the output of `uname -m` command (which is typically either `x86_64` or `aarch64`), and later uses the variable to set the path to the static library.